+++ crud/update.lua
@@ -4,6 +4,7 @@ local vshard = require('vshard')
 
 local call = require('crud.common.call')
 local utils = require('crud.common.utils')
+local sharding = require('crud.common.sharding')
 local dev_checks = require('crud.common.dev_checks')
 
 local UpdateError = errors.new_class('Update',  {capture_stack = false})
@@ -45,6 +46,10 @@ end
 -- @tparam ?number opts.timeout
 --  Function call timeout
 --
+-- @tparam ?number opts.bucket_id
+--  Bucket ID
+--  (by default, it's vshard.router.bucket_id_strcrc32 of primary key)
+--
 -- @return[1] object
 -- @treturn[2] nil
 -- @treturn[2] table Error description
@@ -52,6 +57,7 @@ end
 function update.call(space_name, key, user_operations, opts)
     checks('string', '?', 'table', {
         timeout = '?number',
+        bucket_id = '?number|cdata',
     })
 
     opts = opts or {}
@@ -71,7 +77,7 @@ function update.call(space_name, key, user_operations, opts)
         return nil, UpdateError:new("Wrong operations are specified: %s", err)
     end
 
-    local bucket_id = vshard.router.bucket_id_strcrc32(key)
+    local bucket_id = sharding.key_get_bucket_id(key, opts.bucket_id)
     local result, err = call.rw_single(
         bucket_id, UPDATE_FUNC_NAME, {space_name, key, operations},
         {timeout=opts.timeout})